requirements.txt
pandas 
numpy 
pymongo
matplotlib
python-dotenv
python-box==6.0.2
ensure==1.0.2
joblib
pyYAML
scikit-learn
-e .
